Enh: refactor data quality code from peslab to quality
